+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+* Add the Extend operator: extends operand tuples with attributes
+resulting from specified computations.
+
 # 0.1.1
 
 * Options passed to Reader::Excel and Bmg.excel are passed, unchanged to Roo::Spreadsheet.